cache: use number of chars across expressions for size limit
Current implementation limits the cache size by number of entries
and is equal to 10k entries. However, the size of each entry may
significantly vary depending on expression complexity. This could
result into unexpected memory usage by this lib for cases when
rich and complex expressions are used.

On the other hand, size of the expression supposed to correlate
with its string representation. In this change, we switch from
limiting the cache size by entries to limiting by total number
of chars across all stored expressions. This supposed to put more
strict boundaries on memory usage for scenarios where simple and
complex regexp are used.