prettifier_test.go: add more tests for verifying invalid durations at step and offset
While at it, use ', ' instead of ',' as delimiter of formatted WITH templates for better readability.
So now `WITH(x=a,y=b)` is formatted as `WITH (x=a, y=b)` instead of `WITH (x=a,y=b)`.

This is a follow-up for e2faea4883dc39f0969ce9a4c4bedc4a86d67c41

Updates https://github.com/VictoriaMetrics/VictoriaMetrics/issues/6736