Properly handle $__interval and $__rate_interval
The $__interval and $__rate_interval must be skipped inside [...], in order to trigger automatic detection
of the lookbehind window at VictoriaMetrics side.

This means that `rate(m[$__rate_interval])` and `rate(m[$__interval])` must be parsed as `rate(m)`.

Converting $__interval and $__rate_interval to 1i doesn't work as expected, since `rate(m[1i])`
takes into account only samples on the interval `(t-step ... t]`, and ignores older samples.
This means that it will return unexpected results for the most of users when this interval
doesn't contain enough samples to calculate the `rate`.

This is a follow-up for ce4fd13b702ad0704b5470f51826c1175e5e9596